🔨modify:  refresh token 추가

+++ src/hooks/comment/useCommentPutDelete.ts
@@ -1,8 +1,8 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import axios from 'axios';
-import { toast } from 'react-toastify';
 
 import { getCookies } from '@/utils/getCookies';
+import { postCookies } from '@/utils/postCookies';
 
 const useCommentPutDelete = (
   id: number,
@@ -10,42 +10,65 @@ const useCommentPutDelete = (
   postid: number,
   content?: string
 ) => {
+  const token = getCookies('jwt');
   const queryClient = useQueryClient();
 
   const { mutate: mutateDelete } = useMutation({
     mutationFn: async () => {
+      if (!id && !token) {
+        return;
+      }
+
       return await axios.delete(
         `${process.env.BASE_URL}/comment/${id}?postType=${type}`,
         {
-          headers: { Authorization: getCookies('jwt') },
+          headers: { Authorization: token },
         }
       );
     },
-    onSuccess: () => {
+    onSuccess: (data) => {
+      if (data) {
+        if (data.status === 202 && !data.data.success) {
+          postCookies({
+            jwt: data.data.data.jwt,
+            memberId: data.data.data.memberId,
+          });
+        }
+      }
+
       queryClient.invalidateQueries({ queryKey: [`comment-${postid}`] });
     },
-    onError: () => {
-      toast.error('에러 발생');
+    onError: (error) => {
+      console.error(error);
     },
   });
 
   const { mutate: mutatePut, isSuccess } = useMutation({
     mutationFn: async () => {
-      if (!content) return null;
+      if (!content && !id && !token) return;
 
       return await axios.put(
         `${process.env.BASE_URL}/comment/${id}?postType=${type}`,
         { content: content },
         {
-          headers: { Authorization: getCookies('jwt') },
+          headers: { Authorization: token },
         }
       );
     },
-    onSuccess: () => {
+    onSuccess: (data) => {
+      if (data) {
+        if (data.status === 202 && !data.data.success) {
+          postCookies({
+            jwt: data.data.data.jwt,
+            memberId: data.data.data.memberId,
+          });
+        }
+      }
+
       queryClient.invalidateQueries({ queryKey: [`comment-${postid}`] });
     },
-    onError: () => {
-      toast.error('에러 발생');
+    onError: (error) => {
+      console.error(error);
     },
   });
